Girder disassembly changes
Current situation:
Wrench to girder (unsecured) -> secure girder
Wrench to girder (secured) -> disassemble girder
Crowbar to girder (secured) -> unsecure girder

This PR changes it to:
Wrench to girder (unsecured) -> secure girder (same as before)
Wrench to girder (secured) -> unsecure girder
Welder to girder (unsecured) -> disassemble girder

This is more consistent with nearly every other frame in the game, which have to be first unsecured with a wrench and then disassembled with a welder.

+++ code/game/objects/structures/girders.dm
@@ -4,48 +4,55 @@
 	anchored = 1
 	density = 1
 	layer = 2
-	var/state = 0
+	var/state = 0 //state 0 = normal girder, state 1 = no idea, state 2 = reinforced girder, state 3 = displaced/dislodged girder
 	var/girderpasschance = 20 // percentage chance that a projectile passes through the girder.
 
 /obj/structure/girder/attackby(obj/item/W as obj, mob/user as mob, params)
 	add_fingerprint(user)
-	if(istype(W, /obj/item/weapon/wrench) && state == 0)
-		if(anchored && !istype(src,/obj/structure/girder/displaced))
-			playsound(src.loc, 'sound/items/Ratchet.ogg', 100, 1)
-			user << "<span class='notice'>Now disassembling the girder...</span>"
-			if(do_after(user,40))
-				if(!src.loc)
-					return
-				user << "<span class='notice'>You dissasembled the girder!</span>"
-				new /obj/item/stack/sheet/metal(get_turf(src))
-				qdel(src)
-		else if(!anchored)
-			if (!istype(src.loc, /turf/simulated/floor))
-				usr << "<span class='danger'>A floor must be present to secure the girder!</span>"
-				return
-			playsound(src.loc, 'sound/items/Ratchet.ogg', 100, 1)
-			user << "<span class='notice'>Now securing the girder...</span>"
+	if(istype(W, /obj/item/weapon/weldingtool) && state == 3)
+		var/obj/item/weapon/weldingtool/WT = W
+		if(WT.remove_fuel(0,user))
+			playsound(src.loc, 'sound/items/Welder2.ogg', 50, 1)
+			user.visible_message("<span class='warning'>[user] disassembles the girder.</span>", \
+								"You start to disassemble the girder...", "You hear welding and clanking.")
 			if(do_after(user, 40))
+				if( !WT.isOn() )
+					return
 				if(!src.loc)
 					return
-				user << "<span class='notice'>You secured the girder!</span>"
-				var/obj/structure/girder/G = new (loc)
-				transfer_fingerprints_to(G)
+				user << "<span class='notice'>You disassemble the girder.</span>"
+				var/obj/item/stack/sheet/metal/M = new (loc, 2)
+				transfer_fingerprints_to(M)
 				qdel(src)
 
+	else if(istype(W, /obj/item/weapon/wrench) && state == 3 && 4)
+		if (!istype(src.loc, /turf/simulated/floor))
+			usr << "<span class='warning'>A floor must be present to secure the girder!</span>"
+			return
+		playsound(src.loc, 'sound/items/Ratchet.ogg', 100, 1)
+		user << "<span class='notice'>You start securing the girder...</span>"
+		var oldloc = src.loc
+		if(do_after(user, 40))
+			if(src.loc!=oldloc)
+				return
+			user << "<span class='notice'>You secure the girder.</span>"
+			var/obj/structure/girder/G = new (loc)
+			transfer_fingerprints_to(G)
+			qdel(src)
+
 	else if(istype(W, /obj/item/weapon/gun/energy/plasmacutter))
-		user << "<span class='notice'>Now slicing apart the girder...</span>"
+		user << "<span class='notice'>You start slicing apart the girder...</span>"
 		if(do_after(user,30))
 			if(!src.loc)
 				return
-			user << "<span class='notice'>You slice apart the girder!</span>"
+			user << "<span class='notice'>You slice apart the girder.</span>"
 			new /obj/item/stack/sheet/metal(get_turf(src))
 			qdel(src)
 
 	else if(istype(W, /obj/item/weapon/pickaxe/drill/jackhammer))
 		var/obj/item/weapon/pickaxe/drill/jackhammer/D = W
 		if(!D.bcell.use(D.drillcost))
-			user << "<span class='notice'>Your [D.name] doesn't have enough power to break through the [name].</span>"
+			user << "<span class='warning'>Your [D.name] doesn't have enough power to break through the [name]!</span>"
 			return
 		D.update_icon()
 		user << "<span class='notice'>You smash through the girder!</span>"
@@ -55,38 +62,38 @@
 
 	else if(istype(W, /obj/item/weapon/screwdriver) && state == 2 && istype(src,/obj/structure/girder/reinforced))
 		playsound(src.loc, 'sound/items/Screwdriver.ogg', 100, 1)
-		user << "<span class='notice'>Now unsecuring support struts...</span>"
+		user << "<span class='notice'>You start unsecuring support struts...</span>"
 		if(do_after(user,40))
 			if(!src.loc)
 				return
-			user << "<span class='notice'>You unsecured the support struts!</span>"
+			user << "<span class='notice'>You unsecure the support struts.</span>"
 			state = 1
 
 	else if(istype(W, /obj/item/weapon/wirecutters) && istype(src,/obj/structure/girder/reinforced) && state == 1)
 		playsound(src.loc, 'sound/items/Wirecutter.ogg', 100, 1)
-		user << "<span class='notice'>Now removing support struts...</span>"
+		user << "<span class='notice'>You start removing support struts...</span>"
 		if(do_after(user,40))
 			if(!src.loc)
 				return
-			user << "<span class='notice'>You removed the support struts!</span>"
+			user << "<span class='notice'>You remove the support struts.</span>"
 			var/obj/structure/girder/G = new (loc)
 			transfer_fingerprints_to(G)
 			qdel(src)
 
-	else if(istype(W, /obj/item/weapon/crowbar) && state == 0 && anchored )
-		playsound(src.loc, 'sound/items/Crowbar.ogg', 100, 1)
-		user << "<span class='notice'>Now dislodging the girder...</span>"
+	else if(istype(W, /obj/item/weapon/wrench) && state == 0 && anchored )
+		playsound(src.loc, 'sound/items/Ratchet.ogg', 100, 1)
+		user << "<span class='notice'>Now unsecuring the girder...</span>"
 		if(do_after(user, 40))
 			if(!src.loc)
 				return
-			user << "<span class='notice'>You dislodged the girder!</span>"
+			user << "<span class='notice'>You unsecure the girder.</span>"
 			var/obj/structure/girder/displaced/D = new (loc)
 			transfer_fingerprints_to(D)
 			qdel(src)
 
 	else if(istype(W, /obj/item/stack/sheet))
 		if (!istype(src.loc, /turf/simulated/floor))
-			usr << "<span class='danger'>The girder is too unstable to build anything!</span>"
+			usr << "<span class='warning'>A floor must be present to build a false wall!</span>"
 			return
 
 		var/obj/item/stack/sheet/S = W
@@ -95,23 +102,23 @@
 			if(/obj/item/stack/sheet/metal, /obj/item/stack/sheet/metal/cyborg)
 				if(!anchored)
 					if (S.use(2))
-						user << "<span class='notice'>You create a false wall! Push on it to open or close the passage.</span>"
+						user << "<span class='notice'>You create a false wall. Push on it to open or close the passage.</span>"
 						var/obj/structure/falsewall/F = new (loc)
 						transfer_fingerprints_to(F)
 						qdel(src)
 					else
-						user << "<span class='warning'>You need two sheets of metal to create a false wall.</span>"
+						user << "<span class='warning'>You need two sheets of metal to create a false wall!</span>"
 						return
 				else
 					if(S.get_amount() < 2)
-						user << "<span class='warning'>You need two sheets of metal to finish a wall.</span>"
+						user << "<span class='warning'>You need two sheets of metal to finish a wall!</span>"
 						return
 					user << "<span class='notice'>Now adding plating...</span>"
 					if (do_after(user, 40))
 						if(loc == null || S.get_amount() < 2)
 							return
 						S.use(2)
-						user << "<span class='notice'>You added the plating!</span>"
+						user << "<span class='notice'>You add the plating.</span>"
 						var/turf/Tsrc = get_turf(src)
 						Tsrc.ChangeTurf(/turf/simulated/wall)
 						for(var/turf/simulated/wall/X in Tsrc.loc)
@@ -123,22 +130,22 @@
 			if(/obj/item/stack/sheet/plasteel)
 				if(!anchored)
 					if(S.amount < 2)
-						user << "<span class='warning'>You need at least two sheets to create a false wall.</span>"
+						user << "<span class='warning'>You need at least two sheets to create a false wall!</span>"
 						return
 					S.use(2)
-					user << "<span class='notice'>You create a false wall! Push on it to open or close the passage.</span>"
+					user << "<span class='notice'>You create a false wall. Push on it to open or close the passage.</span>"
 					var/obj/structure/falsewall/reinforced/FW = new (loc)
 					transfer_fingerprints_to(FW)
 					qdel(src)
 				else
 					if (src.icon_state == "reinforced") //I cant believe someone would actually write this line of code...
 						if(S.amount < 1) return ..()
-						user << "<span class='notice'>Now finalising reinforced wall...</span>"
+						user << "<span class='notice'>You start finalizing the reinforced wall...</span>"
 						if(do_after(user, 50))
 							if(!src.loc || !S || S.amount < 1)
 								return
 							S.use(1)
-							user << "<span class='notice'>Wall fully reinforced!</span>"
+							user << "<span class='notice'>You fully reinforce the wall.</span>"
 							var/turf/Tsrc = get_turf(src)
 							Tsrc.ChangeTurf(/turf/simulated/wall/r_wall)
 							for(var/turf/simulated/wall/r_wall/X in Tsrc.loc)
@@ -147,12 +154,12 @@
 						return
 					else
 						if(S.amount < 1) return ..()
-						user << "<span class='notice'>Now reinforcing girders...</span>"
+						user << "<span class='notice'>You start reinforcing the girder...</span>"
 						if (do_after(user,60))
 							if(!src.loc || !S || S.amount < 1)
 								return
 							S.use(1)
-							user << "<span class='notice'>Girders reinforced!</span>"
+							user << "<span class='notice'>You reinforce the girder.</span>"
 							var/obj/structure/girder/reinforced/R = new (loc)
 							transfer_fingerprints_to(R)
 							qdel(src)
@@ -236,6 +243,7 @@
 	name = "displaced girder"
 	icon_state = "displaced"
 	anchored = 0
+	state = 3
 	girderpasschance = 25
 
 /obj/structure/girder/reinforced
@@ -271,23 +279,28 @@
 	layer = 2
 
 /obj/structure/cultgirder/attackby(obj/item/W as obj, mob/user as mob, params)
-	if(istype(W, /obj/item/weapon/wrench))
-		playsound(src.loc, 'sound/items/Ratchet.ogg', 100, 1)
-		user << "<span class='notice'>Now disassembling the girder...</span>"
-		if(do_after(user,40))
-			user << "<span class='notice'>You dissasembled the girder!</span>"
-			var/obj/effect/decal/remains/human/R = new (get_turf(src))
-			transfer_fingerprints_to(R)
-			qdel(src)
+	add_fingerprint(user)
+	if(istype(W, /obj/item/weapon/weldingtool))
+		var/obj/item/weapon/weldingtool/WT = W
+		if(WT.remove_fuel(0,user))
+			playsound(src.loc, 'sound/items/Welder2.ogg', 50, 1)
+			user.visible_message("<span class='warning'>[user] disassembles the girder.</span>", \
+								"You start to disassemble the girder...", "You hear welding and clanking.")
+			if(do_after(user, 40))
+				if( !WT.isOn() )
+					return
+				user << "<span class='notice'>You disassemble the girder.</span>"
+				var/obj/effect/decal/remains/human/R = new (get_turf(src))
+				transfer_fingerprints_to(R)
+				qdel(src)
 
 	else if(istype(W, /obj/item/weapon/gun/energy/plasmacutter))
-		user << "<span class='notice'>Now slicing apart the girder...</span>"
+		user << "<span class='notice'>You start slicing apart the girder...</span>"
 		if(do_after(user,30))
-			user << "<span class='notice'>You slice apart the girder!</span>"
+			user << "<span class='notice'>You slice apart the girder.</span>"
 			var/obj/effect/decal/remains/human/R = new (get_turf(src))
 			transfer_fingerprints_to(R)
 			qdel(src)
-
 
 	else if(istype(W, /obj/item/weapon/pickaxe/drill/jackhammer))
 		var/obj/item/weapon/pickaxe/drill/jackhammer/D = W
